Extract Edit/Delete functionality into EditTripViewModel
Shift Attendees and InviteUsers into Trip directory
