+++ TourMate/TourMate/Frontend/Trip/Views/EditTripView/EditTripView.swift
@@ -10,10 +10,10 @@ import SwiftUI
 struct EditTripView: View {
     @Environment(\.dismiss) var dismiss
 
-    @StateObject var viewModel: TripViewModel
+    @StateObject var viewModel: EditTripViewModel
 
     init(trip: Trip) {
-        self._viewModel = StateObject(wrappedValue: TripViewModel(trip: trip))
+        self._viewModel = StateObject(wrappedValue: EditTripViewModel(trip: trip))
     }
 
     var body: some View {
